test-django-celery-beat-timezone/requirements.txt
Django==4.2
django-celery-beat

redis

psycopg  # connection to Postgres database
ipdb

django-debug-toolbar

channels[daphne]

# -e ../../code/sentry-python  # sdk in a sibling folder to this projects folder
sentry-sdk